\name{svc-package}
\alias{svc-package}
\alias{svc}
\docType{package}
\title{
\packageTitle{svc}
}
\description{
\packageDescription{svc}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{svc}

Maintainer: \packageMaintainer{svc}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = svc)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{svc}}
%\section{Documentation Index}{\packageIndices{svc}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
