% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svc.R
\name{svc}
\alias{svc}
\title{Make a Seats-Votes Curve}
\usage{
svc(plans, dvs, swing = svc_unif)
}
\arguments{
\item{plans}{redist_plans object}

\item{dvs}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> column in \code{plans}}

\item{swing}{function to swing voteshares with}
}
\value{
ggplot2 object
}
\description{
Make a Seats-Votes Curve
}
\examples{
library(redist)
data(iowa)
ia <- redist_map(iowa, existing_plan = cd_2010, pop_tol = 0.01)
plans <- redist_smc(ia, 100, silent = TRUE)
plans <- plans \%>\% dplyr::mutate(dvs = group_frac(ia, dem_08, dem_08 + rep_08))
svc(plans, dvs)

}
